BF: Prevent needless call to 7z when there is no archive
Fixes datalad/datalad#5822
Fixes datalad/datalad#4336

The error condition in datalad/datalad#4336 is hard to trigger. It
relied on another bug in this code: broken URL key handling.
(see https://github.com/datalad/datalad/issues/4336#issuecomment-886741012)

There will be no dedicated test for this particular issue, but for URL
key handling in general.

+++ datalad/distributed/ora_remote.py
@@ -188,6 +188,12 @@ class LocalIO(IOBase):
         )
 
     def get_from_archive(self, archive, src, dst, progress_cb):
+        # Upfront check to avoid cryptic error output
+        # https://github.com/datalad/datalad/issues/4336
+        if not self.exists(archive):
+            raise RIARemoteError("archive {arc} does not exist."
+                                 "".format(arc=archive))
+
         # this requires python 3.5
         with open(dst, 'wb') as target_file:
             subprocess.run([
